+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Hi! This is where we document all notable changes, including bug fixes, enhancements, and dependency updates.
 Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](http://semver.org/) format.
 
+## 0.1.17 2025-06-08
+
+### Maintenance
+
+- Updated Supabase, Svelte, SvelteKit, eslint, vitest.
+
 ## 0.1.16 2025-05-25
 
 ### Added